# openapi.yaml
openapi: 3.0.3
info:
  title: Watcher
  description: watch speicify page.css and send notifications
  version: 0.0.1
servers:
  - url: https://watcher.xiantank.xyz/api/v1
    description: server for demo
  - url: http://localhost:3000/api/v1
    description: run local
...
paths:
...
  /crawlers:
    get:
      tags:
        - crawler
      description: crawlers...
      responses:
        '200':
          description: success get list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crawler"
    post:
      tags:
        - crawler
      description: "create user"
      requestBody:
        $ref: '#/components/requestBodies/Crawler'
      responses:
        '201':
          description: create success
        '409':
          description: username exists
        '422':
          description: format error
  /crawlers/{crawlerId}:
    get:
      tags:
        - crawler
      description: crawlers...
      parameters:
        - name: 'crawlerId'
          required: true
          description: the crawler identifier, as crawlerId
          in: path
          schema:
            type: integer
      responses:
        '200':
          description: success get list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crawler"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
...
  requestBodies:
...
    Crawler:
      description: a JSON object containing user data
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Crawler"
  schemas:
...
    Crawler:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        interval_type:
          type: integer
        url:
          type: string
        parser:
          type: string
        parse_attributes:
          type: integer
        target:
          type: string
          nullable: true
        compare_type:
          type: integer
          enum: [1,2]
...
security:
  - cookieAuth: []